+++ src/schemes/fluid/surface_normal_sph.jl
@@ -237,8 +237,6 @@ function calc_curvature!(system::AbstractFluidSystem, neighbor_system::AbstractF
     system_coords = current_coordinates(u_system, system)
     neighbor_system_coords = current_coordinates(u_neighbor_system, neighbor_system)
 
-    set_zero!(correction_factor)
-
     foreach_point_neighbor(system, neighbor_system,
                            system_coords, neighbor_system_coords,
                            semi) do particle, neighbor, pos_diff, distance
@@ -261,11 +259,6 @@ function calc_curvature!(system::AbstractFluidSystem, neighbor_system::AbstractF
         end
     end
 
-    # Eq. 23
-    for particle in each_integrated_particle(system)
-        curvature[particle] /= (correction_factor[particle] + eps())
-    end
-
     return system
 end
 
@@ -280,6 +273,7 @@ function compute_curvature!(system::AbstractFluidSystem,
 
     # Reset surface curvature
     set_zero!(cache.curvature)
+    set_zero!(cache.correction_factor)
 
     @trixi_timeit timer() "compute surface curvature" begin
         foreach_system_wrapped(semi, v_ode,
@@ -295,5 +289,9 @@ function compute_curvature!(system::AbstractFluidSystem,
                             surface_normal_method(neighbor_system))
         end
     end
+
+    # Eq. 23 in Morris 2000
+    cache.curvature ./= (cache.correction_factor .+ eps())
+
     return system
 end